bugfixes for identifiers and integers
identifiers like “s1″ would have the “1” in it highlighted as a number, which looked strange. This was fixed by adding rules for identifiers.

The number “0” would not appear highlighted in layout declarations, as opposed to other numbers. Hex and octal constants in layouts were also not handled, even if it seems like GLSL accepts them. That was fixed by adding rules for octal/hex constants in layout contexts (somehow this fixed the “0” bug too.)